# openapi/imperium-openapi.json
{
  "components": {
    "schemas": {
...
      "AlertSimulationRequest": {
        "properties": {
          "category": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "threshold": {
            "format": "double",
            "type": "number"
          }
        },
        "required": [
          "category",
          "target",
          "threshold"
        ],
        "type": "object"
      },
      "AlertSimulationResponse": {
        "properties": {
          "dedupe_key": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "should_trigger": {
            "type": "boolean"
          }
        },
        "required": [
          "should_trigger",
          "dedupe_key",
          "explanation"
        ],
        "type": "object"
      },
...
      "CreatePlaybookRequest": {
        "properties": {
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "response_steps": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "trigger": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "trigger",
          "response_steps"
        ],
        "type": "object"
      },
      "CreateThesisRequest": {
        "properties": {
          "conviction_percent": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "invalidation_criteria": {
            "type": [
              "string",
              "null"
            ]
          },
          "position": {
            "type": [
              "string",
              "null"
            ]
          },
          "review_date": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "type": "object"
      },
...
      "PlaybookView": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "playbook_id": {
            "type": "string"
          },
          "response_steps": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "trigger": {
            "type": "string"
          }
        },
        "required": [
          "playbook_id",
          "title",
          "trigger",
          "response_steps",
          "enabled",
          "created_at"
        ],
        "type": "object"
      },
...
      "ScenarioRequest": {
        "properties": {
          "fx_shock_percent": {
            "format": "double",
            "type": "number"
          },
          "market_drop_percent": {
            "format": "double",
            "type": "number"
          },
          "rate_spike_bps": {
            "format": "double",
            "type": "number"
          },
          "scenario_name": {
            "type": "string"
          }
        },
        "required": [
          "scenario_name",
          "market_drop_percent",
          "fx_shock_percent",
          "rate_spike_bps"
        ],
        "type": "object"
      },
      "ScenarioResponse": {
        "properties": {
          "estimated_loss": {
            "format": "double",
            "type": "number"
          },
          "hedge_notional": {
            "format": "double",
            "type": "number"
          },
          "liquidity_impact": {
            "format": "double",
            "type": "number"
          },
          "scenario_name": {
            "type": "string"
          }
        },
        "required": [
          "scenario_name",
          "estimated_loss",
          "liquidity_impact",
          "hedge_notional"
        ],
        "type": "object"
      },
...
      "ThesisMutationResponse": {
        "properties": {
          "archived": {
            "type": "boolean"
          },
          "thesis_id": {
            "type": "string"
          }
        },
        "required": [
          "thesis_id",
          "archived"
        ],
        "type": "object"
      },
      "ThesisView": {
        "properties": {
          "conviction_percent": {
            "format": "double",
            "type": "number"
          },
          "created_at": {
            "type": "string"
          },
          "invalidation_criteria": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          },
          "review_date": {
            "type": "string"
          },
          "thesis_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "thesis_id",
          "title",
          "position",
          "conviction_percent",
          "rationale",
          "invalidation_criteria",
          "review_date",
          "created_at"
        ],
        "type": "object"
      },
      "TogglePlaybookRequest": {
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "",
    "license": {
      "identifier": "UNLICENSED",
      "name": "UNLICENSED"
    },
    "title": "imperium-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/imperium/alerts/simulate": {
      "post": {
        "operationId": "simulate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertSimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertSimulationResponse"
                }
              }
            },
            "description": "Alert simulation result"
          }
        },
        "tags": [
          "imperium"
        ]
      }
    },
...
    "/api/v1/imperium/journal/playbooks": {
      "get": {
        "operationId": "playbooks",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaybookView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Playbooks"
          }
        },
        "tags": [
          "imperium"
        ]
      },
      "post": {
        "operationId": "create_playbook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaybookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybookView"
                }
              }
            },
            "description": "Created playbook"
          },
          "400": {
            "description": "Invalid payload"
          }
        },
        "tags": [
          "imperium"
        ]
      }
    },
    "/api/v1/imperium/journal/playbooks/{playbook_id}/toggle": {
      "post": {
        "operationId": "toggle_playbook",
        "parameters": [
          {
            "description": "Playbook id",
            "in": "path",
            "name": "playbook_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TogglePlaybookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybookView"
                }
              }
            },
            "description": "Playbook updated"
          },
          "400": {
            "description": "Invalid payload"
          }
        },
        "tags": [
          "imperium"
        ]
      }
    },
...
    "/api/v1/imperium/journal/theses": {
      "get": {
        "operationId": "theses",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ThesisView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Decision journal theses"
          }
        },
        "tags": [
          "imperium"
        ]
      },
      "post": {
        "operationId": "create_thesis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThesisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisView"
                }
              }
            },
            "description": "Created thesis"
          },
          "400": {
            "description": "Invalid payload"
          }
        },
        "tags": [
          "imperium"
        ]
      }
    },
    "/api/v1/imperium/journal/theses/{thesis_id}/archive": {
      "post": {
        "operationId": "archive_thesis",
        "parameters": [
          {
            "description": "Thesis id",
            "in": "path",
            "name": "thesis_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisMutationResponse"
                }
              }
            },
            "description": "Archived thesis"
          },
          "400": {
            "description": "Invalid thesis id"
          }
        },
        "tags": [
          "imperium"
        ]
      }
    },
...
    "/api/v1/imperium/portfolio/scenario/simulate": {
      "post": {
        "operationId": "simulate_scenario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            },
            "description": "Scenario simulation"
          }
        },
        "tags": [
          "imperium"
        ]
      }
    },
...
  },
...
}